added autorization

+++ src/components/NoteCard.tsx
@@ -5,34 +5,58 @@ import {Note} from '../types/notes';
 import {deleteNote} from '../api/notes';
 import {useRepo} from '@automerge/automerge-repo-react-hooks';
 import RemoveCircleIcon from '@mui/icons-material/RemoveCircle';
+import {useAppDispatch, useAppSelector} from '../hooks/useRedux';
+import {setActiveNote} from '../store/reducers/DirsSlice';
 
 interface NoteCardProps extends Note {
     refetchNotes?: () => void;
     isActive?: boolean;
 }
 
-const NoteCard: React.FC<NoteCardProps> = ({id, title, refetchNotes, automergeUrl, dirId, isActive}) => {
+const NoteCard: React.FC<NoteCardProps> = ({
+    id,
+    title,
+    refetchNotes,
+    automergeUrl,
+    dirId,
+    allowedMethods,
+    defaultAccess,
+}) => {
     const navigate = useNavigate();
 
+    const dispatch = useAppDispatch();
+    const {activeNote} = useAppSelector((store) => store.dirsReducer);
+
     const handleClickOnCard = React.useCallback(() => {
+        dispatch(
+            setActiveNote({
+                id,
+                title,
+                automergeUrl,
+                dirId,
+                allowedMethods,
+                defaultAccess,
+            }),
+        );
         navigate(`/notes/${id}`);
-    }, [id, navigate]);
+    }, [allowedMethods, automergeUrl, defaultAccess, dirId, dispatch, id, navigate, title]);
 
     const repo = useRepo();
     const theme = useTheme();
 
     const activeStyle = {
-        backgroundColor: isActive ? theme.palette.action.hover : 'auto',
+        backgroundColor: activeNote ? theme.palette.action.hover : 'auto',
         cursor: 'pointer',
     };
 
     const handleDeleteClick = React.useCallback(
         async (ev: React.MouseEvent, note: Note) => {
             ev.stopPropagation();
-            await deleteNote({id: note.id}).then(() => {
-                repo.delete(note.automergeUrl);
-                refetchNotes?.();
-            });
+            if (note.allowedMethods.includes('delete'))
+                await deleteNote({id: note.id}).then(() => {
+                    repo.delete(note.automergeUrl);
+                    refetchNotes?.();
+                });
         },
         [refetchNotes, repo],
     );
@@ -46,7 +70,7 @@ const NoteCard: React.FC<NoteCardProps> = ({id, title, refetchNotes, automergeUr
             <Typography variant="body2">{title}</Typography>
             <RemoveCircleIcon
                 fontSize={'small'}
-                onClick={(ev) => handleDeleteClick(ev, {id, title, automergeUrl, dirId})}
+                onClick={(ev) => handleDeleteClick(ev, {id, title, automergeUrl, dirId, allowedMethods, defaultAccess})}
             />
         </ListItem>
     );
